# next/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GChat API",
        "description": "Документация API для проекта GChat",
        "version": "1.0.0"
    },
    "paths": {
...
        "/custom-fields": {
            "get": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Список кастомных полей",
                "operationId": "listCustomFields",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/GetCustomFields_CursorParam"
                    },
                    {
                        "$ref": "#/components/parameters/GetCustomFields_LimitParam"
                    },
                    {
                        "$ref": "#/components/parameters/GetCustomFields_DirectionParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CustomFieldListCursorPaginatedDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Создать кастомное поле",
                "operationId": "createCustomField",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CustomFieldDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                }
            }
        },
        "/custom-fields/{id}": {
            "get": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Получить кастомное поле",
                "operationId": "getCustomField",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/CustomFieldIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CustomFieldDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Обновить кастомное поле",
                "operationId": "updateCustomField",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/CustomFieldIdParam"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CustomFieldDto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomField"
                ],
                "summary": "Удалить кастомное поле",
                "operationId": "deleteCustomField",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/CustomFieldIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomFieldDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "entityType": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "options": {
                        "type": "array",
                        "items": {},
                        "nullable": true
                    },
                    "isRequired": {
                        "type": "boolean"
                    },
                    "integrationId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CustomFieldListCursorPaginatedDto": {
                "description": "Cursor-based paginated response with CustomFieldDto items",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CursorPaginatedDto"
                    },
                    {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/CustomFieldDto"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "CursorPaginatedDto": {
                "description": "Generic cursor-based paginated response wrapper",
                "properties": {
                    "items": {
                        "description": "List of items for the current page",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "nextCursor": {
                        "type": "string",
                        "example": "eyJpZCI6Ij...",
                        "nullable": true
                    },
                    "prevCursor": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreCustomFieldRequest": {
                "required": [
                    "key",
                    "name",
                    "entity_type",
                    "type"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "entity_type": {
                        "type": "string",
                        "maxLength": 64
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 64
                    },
                    "options": {
                        "type": "array",
                        "items": {},
                        "nullable": true
                    },
                    "is_required": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "integration_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateCustomFieldRequest": {
                "properties": {
                    "key": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "entity_type": {
                        "type": "string",
                        "maxLength": 64,
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 64,
                        "nullable": true
                    },
                    "options": {
                        "type": "array",
                        "items": {},
                        "nullable": true
                    },
                    "is_required": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "integration_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Request successful"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Meta"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Meta": {
                "properties": {
                    "next_cursor": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "prev_cursor": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "NotFoundResponse": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            },
...
            "ValidationErrorResponse": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/ApiResponse"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "field1": [
                                                    "The field1 is required."
                                                ],
                                                "field2": [
                                                    "The field2 must be an integer."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "GetCustomFields_CursorParam": {
                "name": "cursor",
                "in": "query",
                "description": "Cursor for pagination",
                "schema": {
                    "type": "string",
                    "nullable": true
                }
            },
            "GetCustomFields_LimitParam": {
                "name": "limit",
                "in": "query",
                "description": "Items per page (1..200)",
                "schema": {
                    "type": "integer",
                    "default": 50
                }
            },
            "GetCustomFields_DirectionParam": {
                "name": "direction",
                "in": "query",
                "description": "Pagination direction",
                "schema": {
                    "type": "string",
                    "default": "down",
                    "enum": [
                        "up",
                        "down"
                    ]
                }
            },
            "CustomFieldIdParam": {
                "name": "id",
                "in": "path",
                "description": "Custom field UUID",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
...
        }
    },
...
}